reafctor: split the transform module and move them to the separate folder
