+++ lib/transform/rot-transformer.js
@@ -1,33 +1,24 @@
 const { Transform } = require('stream')
 
+const {
+  CODE_,
+  isLowerCode,
+  isUpperCode,
+} = require('./helpers.js')
+
 const DIRECTION_ = {
   RIGHT: 'right',
   LEFT: 'left'
 }
 
-const CODE_ = {
-  LOWER_A: 97,
-  LOWER_Z: 122,
-  UPPER_A: 65,
-  UPPER_Z: 90,
-}
-
-function isLowerCode(code) {
-  return code >= CODE_.LOWER_A && code <= CODE_.LOWER_Z
-}
-
-function isUpperCode(code) {
-  return code >= CODE_.UPPER_A && code <= CODE_.UPPER_Z
+function isValidShifting(shifting) {
+  return typeof shifting === 'number'
+    && shifting > 0
+    && shifting % 1 === 0
 }
 
 function isValidDirection(direction) {
   return direction === DIRECTION_.RIGHT || direction === DIRECTION_.LEFT
-}
-
-function isValidSifting(shifting) {
-  return typeof shifting === 'number'
-    && shifting > 0
-    && shifting % 1 === 0
 }
 
 class RotTransformerError extends Error {}
@@ -48,11 +39,11 @@ class RotTransformer extends Transform {
 
   constructor(shifting, direction) {
     super()
-    
-    if (!isValidSifting(shifting)) {
+
+    if (!isValidShifting(shifting)) {
       throw new InvalidShiftingError()
     }
-    
+
     if (!isValidDirection(direction)) {
       throw new InvalidDirectionError(direction)
     }
@@ -64,11 +55,11 @@ class RotTransformer extends Transform {
   get shifting() {
     return this._shifting
   }
-  
+
   get direction() {
     return this._direction
   }
-  
+
   _transform(data, _, done) {
     const transformed = [...data].map((code) => this._transformCode(code))
     done(null, Buffer.from(transformed))
@@ -80,14 +71,14 @@ class RotTransformer extends Transform {
     if (isLowerCode(code)) {
       return this._transformNumber(code - LOWER_A) + LOWER_A
     }
-    
+
     if (isUpperCode(code)) {
       return this._transformNumber(code - UPPER_A) + UPPER_A
     }
-    
+
     return code
   }
-  
+
   _transformNumber(number) {
     if (this._direction === DIRECTION_.RIGHT) {
       return this._shiftNumberRight(number)
@@ -107,60 +98,6 @@ class RotTransformer extends Transform {
 RotTransformer.DIRECTION_ = DIRECTION_
 RotTransformer.RotTransformerError = RotTransformerError
 
-class CaesarEncoder extends RotTransformer {
-  constructor() {
-    super(1, DIRECTION_.RIGHT)
-  }
-}
-
-class CaesarDecoder extends RotTransformer {
-  constructor() {
-    super(1, DIRECTION_.LEFT)
-  }
-}
-
-class Rot8Encoder extends RotTransformer {
-  constructor() {
-    super(8, DIRECTION_.RIGHT)
-  }
-}
-
-class Rot8Decoder extends RotTransformer {
-  constructor() {
-    super(8, DIRECTION_.LEFT)
-  }
-}
-
-class AtbashTransformer extends Transform {
-  _transform(data, _, done) {
-    const transformed = [...data].map((code) => this._transformCode(code))
-    done(null, Buffer.from(transformed))
-  }
-  
-  _transformCode(code) {
-    const { LOWER_A, UPPER_A } = CODE_
-    
-    if (isLowerCode(code)) {
-      return this._transformNumber(code - LOWER_A) + LOWER_A
-    }
-    
-    if (isUpperCode(code)) {
-      return this._transformNumber(code - UPPER_A) + UPPER_A
-    }
-    
-    return code
-  }
-  
-  _transformNumber(number) {
-    return 25 - number
-  }
-}
-
 module.exports = {
-  RotTransformer,
-  CaesarEncoder,
-  CaesarDecoder,
-  Rot8Encoder,
-  Rot8Decoder,
-  AtbashTransformer,
+  RotTransformer
 }
